# src/docs/swagger.yaml
basePath: /api/v1
definitions:
  controller._ResponsePostDetail:
    properties:
      code:
        description: 业务响应状态码
        type: integer
      data:
        $ref: '#/definitions/models.ApiPostDetail'
        description: 数据
      message:
        description: 提示信息
        type: string
    type: object
  controller._ResponsePostList:
    properties:
      code:
        description: 业务响应状态码
        type: integer
      data:
        description: 数据
        items:
          $ref: '#/definitions/models.ApiPostDetail'
        type: array
      message:
        description: 提示信息
        type: string
    type: object
  models.ApiPostDetail:
    properties:
      author_id:
        type: integer
      author_name:
        type: string
      community_id:
        type: integer
      content:
        type: string
      created_time:
        type: string
      id:
        type: integer
      introduction:
        type: string
      name:
        type: string
      status:
        type: integer
      title:
        type: string
      vote_num:
        type: integer
    required:
    - community_id
    - content
    - title
    type: object
host: localhost:12345
info:
  contact:
    email: zsc@zhangsichen.cn
    name: zhangsichen
    url: https://blog.zhangsichen.cn
  description: 论坛项目
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://bluebell.zhangsichen.cn
  title: Bluebell 论坛项目
  version: "1.0"
paths:
  /post/{id}:
    get:
      consumes:
      - application/json
      description: 可获取帖子详情接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      - description: 查询参数
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponsePostDetail'
      security:
      - ApiKeyAuth: []
      summary: 帖子详情接口
      tags:
      - 帖子查询相关接口
  /posts:
    get:
      consumes:
      - application/json
      description: 可按分页参数查询帖子列表接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      - description: 分页页数
        in: query
        name: page
        type: string
      - description: 分页大小
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponsePostList'
      security:
      - ApiKeyAuth: []
      summary: 帖子列表接口
      tags:
      - 帖子查询相关接口
  /posts2:
    get:
      consumes:
      - application/json
      description: 可按社区按时间或分数排序查询帖子列表接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      - description: 社区ID(可以为空,如果为空则检索范围是所有帖子)
        in: query
        name: community_id
        type: integer
      - description: 排序方式
        example: score
        in: query
        name: order
        type: string
      - description: 页数
        example: 1
        in: query
        name: page
        type: integer
      - description: 分页大小
        example: 10
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponsePostList'
      security:
      - ApiKeyAuth: []
      summary: 升级版帖子列表接口
      tags:
      - 帖子查询相关接口
swagger: "2.0"